[refactor/#112] 뷰모델 코드 정리
- HomeUiState 파일 분리
- 코드 정리
